Refactor studio routes and handlers; move logic to studio_handlers.py
- Moved route handlers from studio.py to a new module studio_handlers.py to separate routing from business logic.
- Updated route registration in studio.py to use the new handlers.
- Simplified the downloader thread logic in discovery_helpers.py by using job_manager for job submission.
- Introduced a new module iiif_logic.py for IIIF manifest manipulation, including functions for rewriting image URLs and counting canvases.
